+++ docs/_pages/releases.md
@@ -28,6 +28,7 @@ sidebar:
 * Added a few more assertions on `XDocument` - [#2690](https://github.com/fluentassertions/fluentassertions/pull/2690)
     * `[Not]HaveElementWithValue`
     * `NotHaveElement`
+* Added `HaveMilliseconds` and `NotHaveMilliseconds` for assertions on `DateTime` and `DateTimeOffset` - [#2775](https://github.com/fluentassertions/fluentassertions/pull/2775)
 
 ### Improvements
 * Improve failure message for string assertions when checking for equality - [#2307](https://github.com/fluentassertions/fluentassertions/pull/2307)